# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LeagueVerse Fantasy Football Draft API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues:
    get:
      operationId: listLeagues
      tags: [leagues]
      summary: List all leagues
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"
    post:
      operationId: createLeague
      tags: [leagues]
      summary: Create a new league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeagueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
  /leagues/{leagueId}:
    get:
      operationId: getLeague
      tags: [leagues]
      summary: Get a league by ID
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
        "404":
          description: Not found
    patch:
      operationId: updateLeague
      tags: [leagues]
      summary: Update a league
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeagueUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
    delete:
      operationId: deleteLeague
      tags: [leagues]
      summary: Delete a league
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      parameters:
        - name: position
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: leagueId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
  /players/{playerId}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a player
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /leagues/{leagueId}/picks:
    get:
      operationId: listPicks
      tags: [picks]
      summary: List all picks in a league draft
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pick"
    post:
      operationId: makePick
      tags: [picks]
      summary: Make a draft pick
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PickInput"
      responses:
        "201":
          description: Pick made
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pick"
  /leagues/{leagueId}/picks/{pickId}:
    get:
      operationId: getPick
      tags: [picks]
      summary: Get a specific pick
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
        - name: pickId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pick"
...
  /leagues/{leagueId}/summary:
    get:
      operationId: getLeagueSummary
      tags: [leagues]
      summary: Get league summary with stats and status
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueSummary"
components:
  schemas:
...
    League:
      type: object
      required: [id, name, commissionerName, numTeams, numRounds, draftType, scoringType, status, theme, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        commissionerName:
          type: string
        numTeams:
          type: integer
        numRounds:
          type: integer
        draftType:
          type: string
          enum: [snake, auction, linear]
        scoringType:
          type: string
          enum: [standard, ppr, half_ppr]
        status:
          type: string
          enum: [setup, predraft, drafting, completed]
        theme:
          type: string
          enum: [nfl, wwe, star_wars, marvel, cyberpunk, sports_bar, vegas]
        timerSeconds:
          type: ["integer", "null"]
        createdAt:
          type: string
    LeagueInput:
      type: object
      required: [name, commissionerName, numTeams, numRounds, draftType, scoringType]
      properties:
        name:
          type: string
        commissionerName:
          type: string
        numTeams:
          type: integer
          minimum: 4
          maximum: 20
        numRounds:
          type: integer
          minimum: 10
          maximum: 20
        draftType:
          type: string
          enum: [snake, auction, linear]
        scoringType:
          type: string
          enum: [standard, ppr, half_ppr]
        theme:
          type: string
          enum: [nfl, wwe, star_wars, marvel, cyberpunk, sports_bar, vegas]
        timerSeconds:
          type: integer
    LeagueUpdate:
      type: object
      properties:
        name:
          type: string
        theme:
          type: string
          enum: [nfl, wwe, star_wars, marvel, cyberpunk, sports_bar, vegas]
        status:
          type: string
          enum: [setup, predraft, drafting, completed]
        timerSeconds:
          type: integer
    Team:
      type: object
      required: [id, leagueId, name, ownerName, draftPosition, createdAt]
      properties:
        id:
          type: integer
        leagueId:
          type: integer
        name:
          type: string
        ownerName:
          type: string
        draftPosition:
          type: integer
        logoUrl:
          type: ["string", "null"]
        walkUpSong:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Player:
      type: object
      required: [id, name, position, nflTeam, rank, adp, projectedPoints, byeWeek, status]
      properties:
        id:
          type: integer
        name:
          type: string
        position:
          type: string
          enum: [QB, RB, WR, TE, K, DEF]
        nflTeam:
          type: string
        rank:
          type: integer
        adp:
          type: number
        projectedPoints:
          type: number
        byeWeek:
          type: integer
        status:
          type: string
          enum: [active, injured, questionable, out, ir]
        injuryNote:
          type: ["string", "null"]
        isDrafted:
          type: boolean
    Pick:
      type: object
      required: [id, leagueId, teamId, playerId, round, pickInRound, overallPick, createdAt]
      properties:
        id:
          type: integer
        leagueId:
          type: integer
        teamId:
          type: integer
        playerId:
          type: integer
        round:
          type: integer
        pickInRound:
          type: integer
        overallPick:
          type: integer
        grade:
          type: ["string", "null"]
        gradeExplanation:
          type: ["string", "null"]
        valueScore:
          type: ["number", "null"]
        createdAt:
          type: string
        player:
          $ref: "#/components/schemas/Player"
        team:
          $ref: "#/components/schemas/Team"
    PickInput:
      type: object
      required: [teamId, playerId]
      properties:
        teamId:
          type: integer
        playerId:
          type: integer
...
    TeamGrade:
      type: object
      required: [teamId, teamName, overallGrade, gradeScore, summary, positionalGrades]
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        overallGrade:
          type: string
        gradeScore:
          type: number
        summary:
          type: string
        positionalGrades:
          type: object
          additionalProperties:
            type: string
...
    LeagueSummary:
      type: object
      required: [leagueId, totalPicks, picksRemaining, completionPercent, topPicksByPosition, recentPicks, teamGrades]
      properties:
        leagueId:
          type: integer
        totalPicks:
          type: integer
        picksRemaining:
          type: integer
        completionPercent:
          type: number
        topPicksByPosition:
          type: object
          additionalProperties:
            type: string
        recentPicks:
          type: array
          items:
            $ref: "#/components/schemas/Pick"
        teamGrades:
          type: array
          items:
            $ref: "#/components/schemas/TeamGrade"