# lib/api-zod/src/generated/types/leagueUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LeagueVerse Fantasy Football Draft API
 * OpenAPI spec version: 0.1.0
 */
import type { LeagueUpdateStatus } from './leagueUpdateStatus';
import type { LeagueUpdateTheme } from './leagueUpdateTheme';

export interface LeagueUpdate {
  name?: string;
  theme?: LeagueUpdateTheme;
  status?: LeagueUpdateStatus;
  timerSeconds?: number;
}